[change] RESTs divididos em Commands e Queries
