+++ back/account/src/main/java/com/bantads/account/rest/transaction/TransactionCommand.java
@@ -1,29 +1,25 @@
-package com.bantads.account.rest;
+package com.bantads.account.rest.transaction;
 
-import java.util.ArrayList;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.bantads.account.exceptions.AccountNotFound;
 import com.bantads.account.exceptions.InsufficientFunds;
 import com.bantads.account.lib.JsonResponse;
-import com.bantads.account.lib.Statement;
 import com.bantads.account.model.Account;
 import com.bantads.account.model.TransactionDTO;
 import com.bantads.account.services.AccountServices;
 import com.bantads.account.services.TransactionServices;
 
-@CrossOrigin
 @RestController
-public class TransactionREST {
+@CrossOrigin
+public class TransactionCommand {
     @Autowired
     private AccountServices accServ;
 
@@ -85,25 +81,6 @@ public class TransactionREST {
         } catch (InsufficientFunds e) {
             return new JsonResponse(400, "Saldo insuficiente!!", null);
         }
-    }
-
-    @GetMapping(value = "/accounts/{id}/statement", produces = "application/json")
-    public JsonResponse getStatement(@PathVariable Long id, @RequestParam String from, @RequestParam String to) {
-        try {
-            Long fromTS = Long.parseLong(from);
-            Long toTS = Long.parseLong(to);
-            Account acc = accServ.getAccount(id);
-            ArrayList<TransactionDTO> dtos = serv.getAccountTransactions(acc, fromTS, toTS);
-            Statement statement = new Statement(dtos, fromTS, toTS);
-
-            return new JsonResponse(200, "Extrato", statement);
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-            return new JsonResponse(400, "Revise os dados passados!", null);
-        } catch (AccountNotFound e) {
-            return new JsonResponse(404, "Conta não encontrada!", null);
-        }
-
     }
 
 }
